\name{IlluminaHumanMethylation450kmanifest}
\alias{IlluminaHumanMethylation450kmanifest}
\docType{data}
\title{
  Manifest object for the IlluminaHumanMethylation450k array
}
\description{
  This object contains the array design for Illumina's Human Methylation
  450k microarray.
}
\usage{data(IlluminaHumanMethylation450kmanifest)}
\details{
  This package is based on the file
  \code{HumanMethylation450_15017482_v.1.2.csv}.
}
\format{
  An object of class \code{IlluminaMethylationManifest}
}
\references{
  Bibikova, M. et al. (2011) Genomics.
  High density DNA methylation array with single CpG site resolution.
  \url{http://dx.doi.org/10.1016/j.ygeno.2011.07.007}
}
\seealso{
  The description of the class in the minfi package,
  \code{\link[minfi]{IlluminaMethylationManifest-class}}.
}
\examples{
data(IlluminaHumanMethylation450kmanifest)
}
\keyword{datasets}
